Added some more configurable settings for spam. The first setting checks if the commenter IP address is in the DSBL list @ http://www.dsbl.org. If it is, the user is redirected to a warning page mentioning the used address is blocked. Otherwise, comment gets trough. Default setting is off.
The second option is to prevent a SPAM flood. Users can set the number of seconds between comments in the admin user panel. If a comment is posted before that the script dies because of a possible spam flood. (the fact I had 30 something spam comments in about 30 seconds might have influenced this setting. ;-) )

git-svn-id: http://pixelpost.googlecode.com/svn/trunk@90 ff196f25-2e1a-0410-95ca-3d8fd8d346c5