I changed the behaviour of function create_language_url_from_tag() in line 1136.
1)
PHP may not declare the argv variables (php.ini settings), so I changed $_SERVER['argv'][0] to $_SERVER['QUERY_STRING'] which I hope should always work.

2)
When switching the language a few times in browse or about page, each time the lang=XX variable was added to the URI.
I inserted a preg_replace to avoid that.

git-svn-id: http://pixelpost.googlecode.com/svn/trunk@153 ff196f25-2e1a-0410-95ca-3d8fd8d346c5