I'm writing some code to support folders (one level deep) in the 'addons' folder. I managed to get it all working, php files in a subfolder of addons will be sucked in by Pixelpost.
However, there is a major drawback. Some addons (I haven't seen them yet) might have some support PHP files in a subfolder in addons. Most of us misuse the 'includes' folder for addon support files.

In order to fix this problem I made it like this: the folder has to start with a + sign and end with a + sign.

So any php file in /addons/+myAddon+ will get sucked in, other folders are left alone.

git-svn-id: http://pixelpost.googlecode.com/svn/trunk@359 ff196f25-2e1a-0410-95ca-3d8fd8d346c5